Optimize 3
With two-pointer approach. The commented part is what I propose, not
significantly better than the first solution. Could optimize a bit
with unordered_map instead of map, but still, the overall performance
is worse than the two-pointer approach because the complexity is
of O(n^2) whereas it is O(n) for two-pointer approach.