932 in cpp
The problem requires to find a trick in identifying beautiful =
arithmetic subsequence-free.

Which is not intuitive and obvious. I failed to find, and instead used
a brute-force approach. Turned out to exceed the time limitation.

Then I translated the official solution from python to c++.

The core message is "an arithmetic sequence never starts and ends with
elements of different parity"