3072 in cpp
Proposed multiple methods for solving the problem.
Although valid, they were refused due to time requirement.
Mainly because the complexity was O(n^2) in the first
4 solutions (all cpp files except for sortedArray_upperBound).

I had the idea of using multiset that stores elements in order but
I still needed to parse the set element by element, which is O(n), so
the resulted perf. is still O(n^2).

std::upper_bound uses a binary search in a sorted array (which is built
progressively while executing, so this condition is always true after
insertion). The corresponding complexity is O(log_2 n). According to
the chrono timing, the binary search is roughly one order of magnitude
faster than element-wise parse.

For 50000 elements, the execution time for binary search method is of
millisecond, but the sequential parse takes about 2 seconds.

Interesting