647 second solution
Not the optimal one. Tomorrow update with this hint:If “aba” is a palindrome, is “xabax” a palindrome? Similarly is “xabay” a palindrome?

That is, once we find a palindrome we can increase the number of ret by
more than one and skip to the end of the palindrome to continue the
outer loop